Add config files for nl05, nl06, nl07 and nl08
nl05, nl06, nl07, nl08 are four new Noble nodepool launcher nodes and
they need configs. We give them configs by making direct copies of the
old nl01, nl02, nl03, and nl04 configs. nl01 is copied to nl05, nl02 is
copied to nl06 and so on. The only edits I have made are to update the
comments indicating nl01 is responsible for managing min ready nodes.
That responsibility is now nl05's and the comment is updated to reflect
that in the new config file.

To roll this out my plan is to land this change. Then update the
inventory to add nl06. As soon as that inventory update change begins to
deploy nl02 will go in the emergency file and I will manually shutdown
its launcher. Then nl06 should come up and begin serving the providers
that were running on nl02. I have chosen nl02/nl06 as the first pair as
impact is minimal (only openmetal). Depending on how this goes I may
continue to do one launcher at a time of batch up the remaining three.

Change-Id: I22d35939e9b5ff996f57a3053d43cfc46aae025f