Only pause update_constraints.sh when needed
In [0] a 10 minute pause was added to the propose-update-constraints
playbook in order to resolve issues with new releases not showing up
immediately on pypi. While this seems to have worked well, the pause is
not necessary in those cases where no actual review gets generated.

Move the pause deeper into the script so that is only actually executed
when needed.

[0] 45428f1f9f485d810e3d77db89c13931eaba9029

Change-Id: I3dab786627e230d1418a7975be16eff0ea7d1fd9