Add official-openstack-repo-jobs to openstack-helm-infra
We removed the official-openstack-repo-jobs template
from the openstack-helm-infra project here [1].
We erroneously followed the generic OpenDev doc [2]
instead of obeying the OpenStack specific doc [3]
which says that official-openstack-repo-jobs template must
stay there to keep the repo in sync (including the retirement commit)
with the GitHub mirror.

So now to sync the retirement commit we put the template
back to the openstack-helm-infra project definition.

[1] https://review.opendev.org/c/openstack/project-config/+/945127
[2] https://docs.opendev.org/opendev/infra-manual/latest/drivers.html#retiring-a-project
[3] https://docs.openstack.org/project-team-guide/repository.html#retiring-a-repositor

Change-Id: Ic49d2f4b83692a5e58dd6b2d11639a9d2b3777b0