Move OSA sync to integrated repository
At the moment the only valuable asset of tests repo, which we want to
keep is file synchronization.

With that we want to deprecate current tests repository, as it serves
not other purpose. Existing files sync process we aim to move to
openstack-ansible integrated repository and keep it there for now.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/947626
Change-Id: Ib8b4e651900d6a74d21dd07bdef6867e9f6bf351