Prepare for retirement of RefStack repositories
The OpenInfra Foundation adjusted OpenStack's trademark policies
last year to no longer require use of the RefStack service and
tooling. That coupled with the Interop WG ceasing to exist and lack
of maintainership, means we should retire and clan up these
resources.

Add the noop-jobs template to the four repositories first, so that
their content can be removed and replaced.

Change-Id: I406cc4e87033ee6b9490f7cbb55efd2b33fcf1e4
Depends-On: https://review.opendev.org/c/opendev/system-config/+/947856