Make scripts DCO compliant
The OpenStack project will be switching from the OpenInfra
Individual Contributor License Agreement to the Developer
Certificate of Origin, which requires that commit messages include a
Signed-Off-By trailer. Any commits generated by automated tooling
will need to comply with this requirement as well. Simply including
git commit's -s option on the command line should be sufficient to
meet this requirement.

Change-Id: I46a6318d1257c65d573d289dbd597a2b301c1b97