Update nodeset for test-release-openstack
Currently the test-release-openstack job is using Python 3.8 because it
runs on Ubuntu 20.04 (focal). While project can still support building
the packages on older python version we now have a min python version of
3.10 for installing most OpenStack packages.

This change updates the runtime used by the job for stable/2026.1 and
master to Ubuntu 24.04 (noble) and for stable/2024.2 through
stable/2025.2 to Ubuntu 22.04 (jammy). All older, unmaintained branches
continue to use focal. This is done following the pattern established
for openstack-tox in openstack-zuul-jobs.

Change-Id: I8ba9d954078dce091dec7d2380b3b769ca3fb520
Co-authored-by: Stephen Finucane <stephenfin@redhat.com>