Allow late EOL tagging of trailing projects
Cycle-trailing projects (ie. deployment projects: OpenStack Ansible,
Kolla, Puppet OpenStack) have shifted time schedule for their
coordinated releases and also for transitioning to End of Life.
This is needed so that they can consume all other projects latest
changes before the release or EOL-transition.

This change allows to tag <series>-eol trailing type projects even
if the series is marked as End of Life already, so that we don't have
to wait with marking the series as EOL weeks after all the non-trailing
projects already moved to EOL.

Change-Id: I9bbfba0709457d3ad79fee0875539cd863f52939