Drop lingering ubuntu-xenial jobs
We stopped testing on ubuntu-xenial in bindep, but forgot that
project-config was still using it to check changes to the fallback
packages list.

Also publish-openstack-sphinx-docs and its parent have been
deprecated for a long time and seem to no longer be in use, but rely
on an ubuntu-xenial nodeset which we have ceased providing.

Change-Id: I70ecb4c4f17187efb0e3b79d488c3e397b231769