Disable ruamel automagic line wrapping
Recent releases of ruamel have resulted in weird linewrapping. Ruamel
seems to be aware of this based on the ToDo comment in the diff that
introduced the behavior [0]. Rather than fight this by continuing to
exclude ruamel releases we simply set the width to wrap on to the max
int size. This solution was inspired by this Github comment [1].

This means that we will be responsible for line wrapping on widths
manually. I think that is a reasonable compromise if it gives us more
predictable results.

[0] https://sourceforge.net/p/ruamel-yaml/code/ci/375c4db4c7c5fad61e149fc0690e50229df23ab3/tree/emitter.py?diff=7218ee8b837cb2178b5910b5c0d50ab932a5eaed
[1] https://github.com/iterative/dvc/issues/9397#issuecomment-1534045898

Change-Id: I9a8f831009f4811b3633847094bca7b71169bedb