Make update_constraints.sh work with pyproject.toml
Some projects started to remove python classifiers from setup.cfg and
moved them to pyproject.toml. The update_constraints.sh was not updated
for this and such cases the 'new-release' patches were simply not
generated.
This change checks if setup.cfg contain the python version classifiers
and if not (PYTHON_3_VERSIONS variable is empty) then greps the list of
python versions from pyproject.toml.

Change-Id: I2e24eecd468c4a23f28572e7a8cfc54b1551b153
Signed-off-by: Előd Illés <elod.illes@est.tech>