Fix update_constraints.sh for 'python_version>=' cases
'python_version>=3.10' lines were started to be used in
upper-constraints.txt, but update_constraints.sh script expected to use
explicit versions in the constraint files and because of this some
new-release patches were simply scipped as there were no matches for
'python_version==' in the constraints file.

This patch adds the extra check for 'python_version>=' lines and
updates them if needed.

Change-Id: If770b2e7f2d25495c4b455e40f0d2c04ec7e41f6