Update Jeepyb's Gerrit builds to Gerrit 3.11
This change is being pushed in preparation for our upcoming upgrade to
Gerrit 3.11. Once we are upgraded we will want to ensure that jeepyb
updates trigger the correct Gerrit image build. This chagne should only
be merged after the upgrade is complete and we are confident we are
unlikely to rollback.

Change-Id: Iebff82288cc46fa3848a4a539192e9d639f396ed